win32 build tweaks
git-svn-id: http://svn.freeswitch.org/svn/freeswitch/trunk@5540 d0543943-73ff-0310-b7d9-9358b9ac24b2

+++ src/switch_ivr_play_say.c
@@ -942,10 +942,10 @@ SWITCH_DECLARE(switch_status_t) switch_ivr_play_file(switch_core_session_t *sess
 			olen = llen;
 		}
 
-		write_frame.samples = olen;
+		write_frame.samples = (uint32_t)olen;
 
 		if (asis) {
-			write_frame.datalen = olen;
+			write_frame.datalen = (uint32_t)olen;
 		} else {
 			write_frame.datalen = write_frame.samples * 2;
 		}
